Add outpatient routes and controller; update patient routes and dependencies
