+++ package.json
@@ -26,7 +26,7 @@
   },
   "dependencies": {
     "@prisma/client": "^5.18.0",
-    "body-parser": "^1.20.2",
+    "body-parser": "^1.20.3",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
