+++ src/controllers/OutpatientController.ts
@@ -1,138 +1,125 @@
-import { PrismaClient } from "@prisma/client";
 import { Request, Response } from "express";
+import { PrismaClient } from "@prisma/client";
+
 const prisma = new PrismaClient();
 
-export const createPatient = async (req: Request, res: Response) => {
+export const createOutPatient = async (req: Request, res: Response) => {
   try {
     const {
       nic,
-      age,
       name,
+      age,
       gender,
-      address,
+      livingStatus,
       phone,
-      diagnose,
+      description,
       reason,
-      pressure,
-      weight,
       city,
       stateProvince,
       postalCode,
       country,
       streetAddress,
-      livingStatus,
+      prescriptions,
     } = req.body;
-
-    const patient = await prisma.patient.findUnique({
-      where: {
-        nic,
-      },
-    });
-    if (patient) {
-      return res.status(400).json({ message: "Patient already exists" });
-    }
-
-    const newPatient = await prisma.patient.create({
+    console.log("req.body", req.body);
+    console.log("nic", nic);
+    const newOutPatient = await prisma.outPatientFrom.create({
       data: {
         nic,
-        age,
         name,
+        age,
         gender,
-        livingStatus,
+        prescriptions,
         phone,
+        description,
         reason,
-        streetAddress,
         city,
         stateProvince,
         postalCode,
         country,
+        streetAddress,
+        livingStatus: "active", // or any default value
       },
     });
 
-    res
-      .status(200)
-      .json({ message: "Patient Created Successfully", newPatient });
+    res.status(201).json({
+      message: "OutPatient Created Successfully",
+      newOutPatient,
+    });
   } catch (error: any) {
-    res
-      .status(500)
-      .json({ message: `Error creating Patient: ${error.message}` });
+    res.status(400).json({ message: error.message });
   }
 };
 
-export const getAllPatients = async (req: Request, res: Response) => {
+export const getAllOutPatients = async (req: Request, res: Response) => {
   try {
-    const Patients = await prisma.patient.findMany();
-    res
-      .status(200)
-      .json({ message: "Patients fetched successfully!", Patients });
+    const outPatients = await prisma.outPatientFrom.findMany();
+    res.status(200).json(outPatients);
   } catch (error: any) {
-    res
-      .status(500)
-      .json({ message: `Error getting Patients: ${error.message}` });
-  }
-};
-export const getPatientByNic = async (req: Request, res: Response) => {
-  const { nic } = req.params;
-
-  try {
-    const Patient = await prisma.patient.findUnique({
-      where: {
-        nic,
-      },
-    });
-
-    if (!Patient) {
-      // Respond with a 404 if the patient is not found
-      return res.status(404).json({ message: "Patient not found" });
-    }
-
-    res.status(200).json({ message: "Patient fetched successfully!", Patient });
-  } catch (error: any) {
-    res
-      .status(500)
-      .json({ message: `Error getting Patient: ${error.message}` });
+    res.status(400).json({ message: error.message });
   }
 };
 
-export const deletePatient = async (req: Request, res: Response) => {
-  const { nic } = req.params;
-  if (!nic) {
-    return res.status(400).json({ message: "nic is required" });
-  }
-  try {
-    const deletedPatient = await prisma.patient.delete({
-      where: {
-        nic,
-      },
-    });
-    res.status(200).json({
-      message: "Patient Deleted Successfully",
-    });
-  } catch (error: any) {
-    res
-      .status(500)
-      .json({ message: `Error deleting Patient: ${error.message}` });
-  }
-};
+// export const getOutPatientByNic = async (req: Request, res: Response) => {
+//   try {
+//     const { nic } = req.params;
+//     if (!nic) {
+//       return res.status(400).json({ error: "NIC is required." });
+//     }
+//     const outPatient = await prisma.outPatientFrom.findUnique({
+//       where: {
+//         nic: nic,
+//       },
+//     });
+//     res
+//       .status(200)
+//       .json({ message: "OutPatient fetched successfully!", outPatient });
+//   } catch (error: any) {
+//     res
+//       .status(500)
+//       .json({ message: `Error getting OutPatient: ${error.message}` });
+//   }
+// };
 
-export const updatePatient = async (req: Request, res: Response) => {
-  const { nic } = req.params;
-  if (!nic) {
-    return res.status(400).json({ message: "nic is required" });
-  }
-  try {
-    const updatedPatient = await prisma.patient.update({
-      where: {
-        nic,
-      },
-      data: req.body,
-    });
-    res.status(200).json({
-      message: "Patient Updated Successfully",
-    });
-  } catch (error: any) {
-    res
-      .status(500)
-      .json({ message: `Error updating Patient: ${error.message}` });
-  }
-};
+// export const updateOutPatient = async (req: Request, res: Response) => {
+//   try {
+//     const { nic } = req.params;
+//     if (!nic) {
+//       return res.status(400).json({ error: "NIC is required." });
+//     }
+//     const updatedOutPatient = await prisma.outPatientFrom.update({
+//       where: {
+//         nic,
+//       },
+//       data: req.body,
+//     });
+//     res
+//       .status(200)
+//       .json({ message: "OutPatient updated successfully!", updatedOutPatient });
+//   } catch (error: any) {
+//     res
+//       .status(500)
+//       .json({ message: `Error updating OutPatient: ${error.message}` });
+//   }
+// };
+
+// export const deleteOutPatient = async (req: Request, res: Response) => {
+//   try {
+//     const { nic } = req.params;
+//     if (!nic) {
+//       return res.status(400).json({ error: "NIC is required." });
+//     }
+//     const deletedOutPatient = await prisma.outPatientFrom.delete({
+//       where: {
+//         nic,
+//       },
+//     });
+//     res
+//       .status(200)
+//       .json({ message: "OutPatient deleted successfully!", deletedOutPatient });
+//   } catch (error: any) {
+//     res
+//       .status(500)
+//       .json({ message: `Error deleting OutPatient: ${error.message}` });
+//   }
+// };
